move transcoding stuff to "encode" package
